another try

+++ src/lambda/moneybird-auth-init.js
@@ -18,7 +18,7 @@ exports.handler = async (event, context, callback) => {
     console.log(account)
     console.log("-------------ACOUNTDATA UNFILTERED---------------")
     let authData = {
-      'account_id' : parseInt(account[0].id, 10),
+      'account_id' : parseFloat(account[0].id, 10),
       'account_name' : account[0].name,
       'access_token' : tokens.access_token,
       'refresh_token' : tokens.refresh_token
